requirements.txt
pandas
psycopg
selenium
datetime
time
os
pdfplumber
python-dotenv
